+++ testing/test-common/src/main/kotlin/net/corda/testing/common/internal/ParametersUtilities.kt
@@ -5,12 +5,12 @@ import net.corda.core.node.NotaryInfo
 import java.time.Instant
 
 fun testNetworkParameters(
-        notaries: List<NotaryInfo>,
+        notaries: List<NotaryInfo> = emptyList(),
         minimumPlatformVersion: Int = 1,
         modifiedTime: Instant = Instant.now(),
         maxMessageSize: Int = 10485760,
         // TODO: Make this configurable and consistence across driver, bootstrapper, demobench and NetworkMapServer
-        maxTransactionSize: Int = Int.MAX_VALUE,
+        maxTransactionSize: Int = maxMessageSize,
         epoch: Int = 1
 ): NetworkParameters {
     return NetworkParameters(